refactor pcap analyzer test project structure + add file extracting test
