+++ BruteShark/PcapAnalyzerTest/PcapAnalyzerTest.csproj
@@ -49,6 +49,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AnalyzerTest.cs" />
+    <Compile Include="PasswordsModuleTest.cs" />
+    <Compile Include="FileExtractingModuleTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
